Modified for agpypeline library
