+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM agdrone/drone-base-image:1.2
+FROM agpypeline
 LABEL maintainer="Chris Schnaufer <schnaufer@email.arizona.edu>"
 
 COPY requirements.txt packages.txt /home/extractor/
@@ -26,5 +26,10 @@ RUN [ -s /home/extractor/requirements.txt ] && \
     rm /home/extractor/requirements.txt)
 
 USER extractor
+COPY configuration.py plotclip.py /home/extractor/
 
-COPY *.py /home/extractor/
+USER root
+RUN chmod a+x /home/extractor/plotclip.py
+
+USER extractor
+ENTRYPOINT ["/home/extractor/plotclip.py"]
